Change FPS to Float

+++ factory-ai-vision/EdgeSolution/modules/InferenceModule/api/models.py
@@ -53,5 +53,5 @@ class CameraModel(BaseModel):
 
 class CamerasModel(BaseModel):
     lva_mode: Literal["http", "grpc"]
-    fps: int
+    fps: float
     cameras: List[CameraModel]
